refactor loop animation component loader
