+++ packages/engine/src/common/constants/PrefabFunctionType.ts
@@ -9,6 +9,7 @@ export type ComponentUpdateFunction = (entity: Entity) => void
 export type ComponentShouldDeserializeFunction = () => boolean
 export type ComponentPrepareForGLTFExportFunction = (object: Object3D) => void
 
+/** @deprecated */
 export type SceneLoaderType = {
   /**
    * @deprecated An override function to specify custom transformations from ECS to JSON
